+++ pytket/tests/circuit_test.py
@@ -31,6 +31,7 @@ from pytket.circuit import (  # type: ignore
     PauliExpBox,
     QControlBox,
     PhasePolyBox,
+    ToffoliBox,
     CustomGateDef,
     CustomGate,
     Qubit,
@@ -403,6 +404,15 @@ def test_boxes() -> None:
     boxes = (cbox, mbox, u2qbox, u3qbox, ebox, pbox, qcbox)
     assert all(box == box for box in boxes)
     assert all(isinstance(box, Op) for box in boxes)
+
+    permutation = {(0, 0): (1, 1), (1, 1): (0, 0)}
+    tb = ToffoliBox(2, permutation)
+    assert tb.type == OpType.ToffoliBox
+    unitary = tb.get_circuit().get_unitary()
+    comparison = np.asarray([[0, 0, 0, 1], [0, 1, 0, 0], [0, 0, 1, 0], [1, 0, 0, 0]])
+    assert np.allclose(unitary, comparison)
+    d.add_toffolibox(tb, [0, 1])
+    assert d.n_gates == 8
 
 
 def test_u1q_stability() -> None: